refactor: rename variables
