+++ commands/list.js
